map and location added
