+++ app/controllers/users_controller.rb
@@ -5,14 +5,7 @@ class UsersController < ApplicationController
     @category = Category.new
     @categories = current_user.categories
   end
-  
-    @users = User.geocoded #returns flats with coordinates
 
-    @markers = @users.map do |user|
-      {
-        lat: user.latitude,
-        lng: user.longitude
-      }
-    end
+
 
 end
