+++ app/javascript/packs/application.js
@@ -1,30 +1,8 @@
 import "bootstrap";
-import { loadDynamicBannerText } from '../components/banner';
-loadDynamicBannerText();
-
-import 'mapbox-gl/dist/mapbox-gl.css'; // <-- you need to uncomment the stylesheet_pack_tag in the layout!
-
+// import { loadDynamicBannerText } from '../components/banner';
+// loadDynamicBannerText();
+import 'mapbox-gl/dist/mapbox-gl.css';
 import { initMapbox } from '../plugins/init_mapbox';
 
 initMapbox();
 
-if (mapElement) {
-  // [ ... ]
-  const markers = JSON.parse(mapElement.dataset.markers);
-  markers.forEach((marker) => {
-    new mapboxgl.Marker()
-      .setLngLat([ marker.lng, marker.lat ])
-      .addTo(map);
-  });
-}
-
-const fitMapToMarkers = (map, markers) => {
-  const bounds = new mapboxgl.LngLatBounds();
-  markers.forEach(marker => bounds.extend([ marker.lng, marker.lat ]));
-  map.fitBounds(bounds, { padding: 70, maxZoom: 15, duration: 0 });
-};
-
-if (mapElement) {
-  // [ ... ]
-  fitMapToMarkers(map, markers);
-}
