+++ app/javascript/plugins/init_atuocomplete.js
@@ -1,0 +1,8 @@
+import places from 'places.js';
+
+const initAutocomplete = () => {
+  const locationInput = document.getElementById('user_address');
+  if (locationInput) {
+    places({ container: locationInput });
+  }
+};
