--- app/javascript/plugins/init_geoder.js
@@ -1,13 +1,0 @@
-    if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(
-            // Success function
-            showPosition,
-            // Error function
-            null,
-            // Options. See MDN for details.
-            {
-               enableHighAccuracy: true,
-               timeout: 5000,
-               maximumAge: 0
-            });
-    }
