+++ app/models/user.rb
@@ -12,7 +12,7 @@ class User < ApplicationRecord
   validates :age, presence: true,  numericality: { only_integer: true }
   has_one_attached :photo
 
-    geocoded_by :address
-  after_validation :geocode, if: :will_save_change_to_address?
+    geocoded_by :location
+  after_validation :geocode, if: :will_save_change_to_location?
 
 end
