+++ app/views/categories/index.html.erb
@@ -14,6 +14,12 @@
         <i class="far fa-plus-square"></i>
       <% end %>
     </div>
+    <div
+  id="map"
+  style="width: 100%;
+  height: 600px;"
+  data-markers="<%= @markers.to_json %>"
+  data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
   </div>
 </div>
   <% end %>
