+++ app/views/categories/show.html.erb
@@ -8,13 +8,7 @@
     <h2><%= @category.name  %></h2>
     <p><%= @category.description %></p>
     <p><%= @category.price %></p>
-    <div
-  id="map"
-  style="width: 100%;
-  height: 600px;"
-  data-markers="<%= @markers.to_json %>"
-  data-mapbox-api-key="<%= ENV['pk.eyJ1Ijoic2hvMTAiLCJhIjoiY2s2YnF1cDB5MTRjbDNucWp1YXI2dWR2YyJ9.2kk83ZNl4Sae_0x8-5dc0g'
-] %>">current_user.location</div>
+
   </div>
 </div>
