+++ config/initializers/geocoder.rb
@@ -16,7 +16,7 @@ Geocoder.configure(
   # supports SocketError and Timeout::Error
   # always_raise: [],
 
-units: :km,
+  units: :km,
 
   # Calculation options
   # units: :mi,                 # :km for kilometers or :mi for miles
