--- db/migrate/20200227035348_add_coordinates_to_users.rb
@@ -1,6 +1,0 @@
-class AddCoordinatesToUsers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :users, :latitude, :float
-    add_column :users, :longitude, :float
-  end
-end
