+++ package.json
@@ -2,10 +2,12 @@
   "name": "Rails_Rent_Family",
   "private": true,
   "dependencies": {
+    "@mapbox/mapbox-gl-geocoder": "^4.5.1",
     "@rails/webpacker": "4.2.2",
     "bootstrap": "^4.4.1",
     "jquery": "^3.4.1",
     "mapbox-gl": "^1.8.1",
+    "places.js": "^1.18.1",
     "popper.js": "^1.16.1",
     "typed.js": "^2.0.11"
   },
